--- packages/berlin/src/components/newOptionCard/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NewOptionCard';
