+++ packages/berlin/src/pages/Cycle.tsx
@@ -5,7 +5,7 @@ import { useNavigate, useParams } from 'react-router-dom';
 import toast from 'react-hot-toast';
 
 // API
-import { GetCycleResponse, fetchCycle, fetchUserVotes, postVotes } from 'api';
+import { GetCycleResponse, QuestionOption, fetchCycle, fetchUserVotes, postVotes } from 'api';
 
 // Hooks
 import useCountdown from '../hooks/useCountdown';
@@ -30,7 +30,10 @@ import { Title } from '../components/typography/Title.styled';
 import BackButton from '../components/backButton';
 import Button from '../components/button';
 import CycleColumns from '../components/cycleColumns';
-import NewOptionCard from '../components/newOptionCard';
+import OptionCard from '../components/optionCard';
+
+type Order = 'asc' | 'desc';
+type LocalUserVotes = ResponseUserVotesType | { optionId: string; numOfVotes: number }[];
 
 const initialHearts = 20;
 
@@ -54,10 +57,8 @@ function Cycle() {
   const { avaliableHearts, setAvaliableHearts } = useAppStore((state) => state);
   const [startAt, setStartAt] = useState<string | null>(null);
   const [endAt, setEndAt] = useState<string | null>(null);
-  const [localUserVotes, setLocalUserVotes] = useState<
-    ResponseUserVotesType | { optionId: string; numOfVotes: number }[]
-  >([]);
-  const [sorting, setSorting] = useState<{ column: string; order: 'desc' | 'asc' }>({
+  const [localUserVotes, setLocalUserVotes] = useState<LocalUserVotes>([]);
+  const [sorting, setSorting] = useState<{ column: string; order: Order }>({
     column: 'pluralityScore',
     order: 'desc',
   });
@@ -141,31 +142,53 @@ function Cycle() {
     }
   };
 
+  const sortByAuthor = (a: QuestionOption, b: QuestionOption, order: Order) => {
+    const usernameA = a.user.username.toUpperCase();
+    const usernameB = b.user.username.toUpperCase();
+    return order === 'desc'
+      ? usernameB.localeCompare(usernameA)
+      : usernameA.localeCompare(usernameB);
+  };
+
+  const sortByAffiliation = (a: QuestionOption, b: QuestionOption, order: Order) => {
+    const affiliationA = a.user.group.name.toUpperCase();
+    const affiliationB = b.user.group.name.toUpperCase();
+    return order === 'desc'
+      ? affiliationB.localeCompare(affiliationA)
+      : affiliationA.localeCompare(affiliationB);
+  };
+
+  const sortByNumOfVotes = (
+    a: QuestionOption,
+    b: QuestionOption,
+    order: Order,
+    localUserVotes: LocalUserVotes,
+  ) => {
+    const votesA = localUserVotes.find((vote) => vote.optionId === a.id)?.numOfVotes || 0;
+    const votesB = localUserVotes.find((vote) => vote.optionId === b.id)?.numOfVotes || 0;
+    return order === 'desc' ? votesB - votesA : votesA - votesB;
+  };
+
+  const sortByVoteScore = (a: QuestionOption, b: QuestionOption, order: Order) => {
+    return order === 'desc' ? b.voteScore - a.voteScore : a.voteScore - b.voteScore;
+  };
+
   const sortedOptions = useMemo(() => {
     const { column, order } = sorting;
     const sorted = [...(currentCycle?.questionOptions ?? [])].sort((a, b) => {
-      if (column === 'author') {
-        const usernameA = a.user.username.toUpperCase();
-        const usernameB = b.user.username.toUpperCase();
-        return order === 'desc'
-          ? usernameB.localeCompare(usernameA)
-          : usernameA.localeCompare(usernameB);
-      } else if (column === 'affiliation') {
-        const affiliationA = a.user.group.name.toUpperCase();
-        const affiliationB = b.user.group.name.toUpperCase();
-        return order === 'desc'
-          ? affiliationB.localeCompare(affiliationA)
-          : affiliationA.localeCompare(affiliationB);
-      } else if (column === 'numOfVotes') {
-        const votesA = localUserVotes.find((vote) => vote.optionId === a.id)?.numOfVotes || 0;
-        const votesB = localUserVotes.find((vote) => vote.optionId === b.id)?.numOfVotes || 0;
-        return order === 'desc' ? votesB - votesA : votesA - votesB;
-      } else {
-        return order === 'desc' ? b.voteScore - a.voteScore : a.voteScore - b.voteScore;
+      switch (column) {
+        case 'author':
+          return sortByAuthor(a, b, order);
+        case 'affiliation':
+          return sortByAffiliation(a, b, order);
+        case 'numOfVotes':
+          return sortByNumOfVotes(a, b, order, localUserVotes);
+        default:
+          return sortByVoteScore(a, b, order);
       }
     });
     return sorted;
-  }, [currentCycle?.questionOptions, sorting]);
+  }, [currentCycle?.questionOptions, localUserVotes, sorting]);
 
   const handleColumnClick = (column: string) => {
     setSorting((prevSorting) => ({
@@ -217,7 +240,7 @@ function Cycle() {
             const userVote = localUserVotes.find((vote) => vote.optionId === option.id);
             const numOfVotes = userVote ? userVote.numOfVotes : 0;
             return (
-              <NewOptionCard
+              <OptionCard
                 key={option.id}
                 option={option}
                 numOfVotes={numOfVotes}
